requirements.txt
networkx==2.8.8
numpy==1.25.0
openai==0.27.8
PyYAML==6.0.1